mingw-w64-crt/stdio/mingw_pformat.c: Perform 64-bit comparison on the mantissa of a floating point number. This fixes #459.
Reference: https://sourceforge.net/p/mingw-w64/bugs/459/

+++ mingw-w64-crt/stdio/mingw_pformat.c
@@ -2112,8 +2112,8 @@ void __pformat_emit_xfloat( __pformat_fpreg_t value, __pformat_t *stream )
   {
     /* taking the rightmost digit in each pass...
      */
-    int c = value.__pformat_fpreg_mantissa & 0xF;
-    if( c == (int) value.__pformat_fpreg_mantissa)
+    unsigned c = value.__pformat_fpreg_mantissa & 0xF;
+    if( c == value.__pformat_fpreg_mantissa)
     {
       /* inserting the radix point, when we reach the last,
        * (i.e. the most significant digit), unless we found no
